+++ src/Schema/SanitizrTuple.php
@@ -4,13 +4,16 @@ namespace Nebalus\Sanitizr\Schema;
 
 use Nebalus\Sanitizr\Exception\SanitizrValidationException;
 
-class SanitizrBatch extends AbstractSanitizrSchema
+class SanitizrTuple extends AbstractSanitizrSchema
 {
     private array $schemas;
     /**
-     * Initializes the batch schema with one or more schema objects.
+     * Initializes the tuple schema with one or more schema objects.
      *
-     * @param AbstractSanitizrSchema ...$schemas One or more schema instances to include in the batch.
+     * A Tuple is a data structure used to describe an array with a fixed length
+     * where each specific position (index) has a specific type.
+     *
+     * @param AbstractSanitizrSchema ...$schemas One or more schema instances to include mapped directly to their positional index.
      */
     public function __construct(
         AbstractSanitizrSchema ...$schemas