+++ src/Types/SanitizrRegexes.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Nebalus\Sanitizr\Types;
-
-final class SanitizrRegexes
-{
-
-}