+++ src/Value/SafeParsedData.php
@@ -48,7 +48,7 @@ readonly class SafeParsedData
     /**
      * @return string
      */
-    public function getErrorMessage(): string
+    public function getErrorMessage(): ?string
     {
         return $this->error;
     }